[ETCM-197] Move all fast sync components to separate package
