+++ src/main/scala/io/iohk/ethereum/blockchain/sync/fast/SyncMetrics.scala
@@ -1,12 +1,12 @@
-package io.iohk.ethereum.blockchain.sync
+package io.iohk.ethereum.blockchain.sync.fast
 
 import java.util.concurrent.atomic.AtomicLong
 
-import io.iohk.ethereum.blockchain.sync.FastSync.SyncState
-import io.iohk.ethereum.metrics.MetricsContainer
 import com.google.common.util.concurrent.AtomicDouble
+import io.iohk.ethereum.blockchain.sync.fast.FastSync.SyncState
+import io.iohk.ethereum.metrics.MetricsContainer
 
-object FastSyncMetrics extends MetricsContainer {
+object SyncMetrics extends MetricsContainer {
 
   private[this] final val PivotBlockNumberGauge =
     metrics.registry.gauge("fastsync.block.pivotBlock.number.gauge", new AtomicDouble(0d))
