+++ src/main/scala/io/iohk/ethereum/db/storage/FastSyncStateStorage.scala
@@ -5,7 +5,7 @@ import java.nio.ByteBuffer
 import akka.util.ByteString
 import boopickle.CompositePickler
 import boopickle.Default._
-import io.iohk.ethereum.blockchain.sync.FastSync._
+import io.iohk.ethereum.blockchain.sync.fast.FastSync._
 import io.iohk.ethereum.db.dataSource.DataSource
 import io.iohk.ethereum.utils.ByteUtils.compactPickledBytes
