+++ src/test/scala/io/iohk/ethereum/blockchain/sync/SyncSchedulerStateSpec.scala
@@ -1,7 +1,7 @@
 package io.iohk.ethereum.blockchain.sync
 
 import akka.util.ByteString
-import io.iohk.ethereum.blockchain.sync.SyncStateScheduler.{SchedulerState, StateNode, StateNodeRequest}
+import io.iohk.ethereum.blockchain.sync.fast.SyncStateScheduler.{SchedulerState, StateNode, StateNodeRequest}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
